requirements.txt
prefect==3.0.3
polars==1.9.0
pyarrow==17.0.0
requests==2.32.3
pydantic==2.8.2
sqlalchemy==2.0.36
psycopg[binary,pool]==3.2.3

langchain==0.3.3
langchain-openai==0.2.2
langgraph==0.2.37
faiss-cpu==1.8.0.post1
tiktoken==0.7.0

streamlit==1.38.0
plotly==5.24.1
python-dotenv==1.0.1